Planar reflection, shadow pass, WebGL2 fixes, and visual polish (#5)
* Fix wave precision decay, animation cancel, Enter key, and integer POST

- Wave shaders: use mod(w*time, 2*PI) to prevent phase precision loss over time
- UniformAnimator: cancel() method drops in-flight animations when a direct set arrives
- REST: accept integer JSON numbers (is_number vs is_number_float) so POST doesn't 404
- Web UI: Enter key in number fields now triggers value post


* Add planar reflection pass for water surface

Renders a Y-flipped pre-pass of non-water scene elements into an FBO,
then samples it in water.frag with normal-perturbed UVs blended against
the cubemap reflection via Fresnel. Desktop only for now.


* Improve planar reflection: clip plane, Fresnel blend, config, distortion uniform

- Clip plane at y=0 in drawable/buoy shaders via gl_ClipDistance[0], enabled
  only during the reflection pass to avoid affecting main draw
- Fresnel-weighted planar vs cubemap blend: planarWeight = 1 - cosTheta so
  grazing angles favour the planar reflection, steep angles favour the cubemap
- Reflection resolution configurable via environment.json "reflection.size"
- Normal perturbation exposed as reflectionDistortion uniform (default 0.03)


* Web UI improvements: nav links, transition control, shader uniform seeding

- Add transition duration control to all web UI pages (buoy, gust, surface)
- Add cross-page navigation links to all web UI pages
- Remove unused LDO page, JSON config, and LODSurface shaders
- Seed UniformState from active GL uniforms on first render callback so
  controls show values on page load without requiring a prior POST
- Prune dead controls from gust.json (causticScale, causticSpeed, fogColor,
  lineColor) and surface.json (ambientStrength) that had no active uniforms


* Update uniforms_min.json with tuned gust and wave settings


* Add PCF shadow map pass for buoy shadow on water surface

Orthographic depth pass from the sun's perspective, sized to the mesh
extent. 3x3 PCF with slope-scaled bias applied to the diffuse term in
water.frag. Shadow map resolution configurable via environment.json.


* Add caustics, waterline effect on buoy, and shadow pass

- PCF shadow map from sun's orthographic view onto water surface
- Waterline foam band and wet-hull darkening on buoy hull
- Wave height evaluated per-fragment with XZ inversion to match
  displaced surface geometry across multiple waves
- GerstnerWave.cpp: apply mod(omega*time, 2PI) to match shader precision fix


* Expose all shader uniforms in web UI with organized pages

- Add vec3 color type to buildControls.js for causticColor, fogColor, deepWaterTint
- buoy.json: add ambientStrength, causticIntensity, waterline/wet params
- gust.json: restructure under mesh_shader group, add all caustic params and foamPower
- surface.json: add reflectionDistortion, fresnelF0, gamma, fogColor, deepWaterTint, depth fade, showMesh


* Modulate waterline band with FBM noise to break up solid halo

Adds waterlineNoise uniform (0=solid ring, 1=fully patchy) that
multiplies the band mask with animated FBM, creating irregular
foam patches instead of a uniform ring around the hull.


* Add hull displacement to buoy via vertex shader noise

Displaces buoy vertices along their normals using two-octave 3D value
noise, with the TBN normal perturbed analytically from the displacement
gradient so lighting responds correctly to the dents. hullDisplacementAmt
uniform (default 0.04) controls amplitude; exposed in web UI.
Works on both desktop and WebGL.


* Add ARCHITECTURE.md with honest assessment of current design


* Fix caustic and waterline noise degeneration over time

As time grows, (FragPos * scale + time * speed) becomes large and
float32 loses precision in fract(), causing the FBM noise to flatten
out. Fix: wrap per-octave inputs to [0,289) in fbm() so fract() always
operates on small values, and wrap neighbor grid coords in valueNoise()
for continuity at the tile boundary. Same fix applied to both water.frag
(caustics, foam) and normal_bump.frag (waterline band noise).


* Add shadow and reflection passes to WebGL build

- Enable ShadowPass and ReflectionPass in Emscripten build
- Work around missing glDrawBuffer/GL_CLAMP_TO_BORDER in WebGL2
- Use fragment discard as software clip plane during reflection pass
- Port shadow PCF and planar reflection to webgl/water.frag
- Add isReflectionPass uniform to webgl/normal_bump.frag and drawable.frag
- Sync uniforms_web.json with uniforms_min.json (fix missing causticScale/causticSpeed and diverged values)
- Add WebGL build instructions to README


* Fix WebGL2/ANGLE compatibility issues

- Use GL_RGB8 sized internal format for 2D textures and cubemap faces
- Fix depth texture format: GL_DEPTH_COMPONENT24 requires GL_UNSIGNED_INT
- Add glDrawBuffers([GL_NONE]) for depth-only shadow FBO completeness in WebGL2
- Add explicit layout(location=) qualifiers to gerstner.vert
- Remove stale 'in vec4 fPosition' from simple.frag that broke wireframe shader linking
- Explicitly set showMesh=1 in drawWireframe to avoid async uniform timing issues
- Initialise ocean wireframe flag explicitly before first draw


* Initialise uninitialized member variables

Uninitialized bool/float/double members behave differently under
Emscripten (WASM zero-initializes memory) vs native builds, causing
subtle bugs. Add default member initializers throughout.

- Drawable: shouldDraw* flags → false
- Camera: isFloating → false
- Ocean, OceanGestalt: lastTime → 0.0
- OceanGestalt: window position/size ints → 0
- TextRenderer: screenWidth/Height → 0.0f
- Biquad: filter coefficients → 0.0f
- SurfAudio: cast RAND_MAX to float to silence implicit conversion warning


* Extend wave startup animation duration and add CLAUDE.md

Wave amplitude animation duration extended from 4s to 30s for a
more gradual build-in on startup. Add CLAUDE.md for Claude Code.


* Silence noisy uniform-not-found warning


* Fix GLM_ENABLE_EXPERIMENTAL in OceanGestalt.cpp

Move the define into the source file instead of patching the
vendored glm header, keeping the submodule clean.


* Consolidate uniforms.json to shared wave paths, use uniforms_min on native

Removes per-shader duplicated wave/uniform keys (mesh_shader.waves[],
normal_shader.waves[], etc.) in favour of the shared waves[] path.
Native build now loads uniforms_min.json to match the web config.


---------

+++ src/core/SurfAudio.cpp
@@ -55,8 +55,8 @@ void SurfAudio::dataCallback(ma_device* device, void* output, const void*, ma_ui
     float steepnessRight = self->steepnessRight.load(std::memory_order_relaxed);
 
     for (ma_uint32 i = 0; i < frameCount; ++i) {
-      float noiseLeft = ((float)rand() / RAND_MAX) * 2.0f - 1.0f;
-      float noiseRight = ((float)rand() / RAND_MAX) * 2.0f - 1.0f;
+      float noiseLeft = ((float)rand() / (float)RAND_MAX) * 2.0f - 1.0f;
+      float noiseRight = ((float)rand() / (float)RAND_MAX) * 2.0f - 1.0f;
 
       float bandNoiseLeft = self->leftFilter.process(noiseLeft);
       float bandNoiseRight = self->rightFilter.process(noiseRight);
